+++ src/Http/Service.php
@@ -6,11 +6,8 @@ namespace Cakasim\Payone\Sdk\Http;
 
 use Cakasim\Payone\Sdk\AbstractService;
 use Psr\Container\ContainerInterface;
-use Psr\Http\Client\ClientInterface;
-use Psr\Http\Message\RequestFactoryInterface;
-use Psr\Http\Message\ResponseFactoryInterface;
-use Psr\Http\Message\StreamFactoryInterface;
-use Psr\Http\Message\UriFactoryInterface;
+use Psr\Http\Message\ServerRequestFactoryInterface;
+use Psr\Http\Message\ServerRequestInterface;
 
 /**
  * The HTTP service.
@@ -21,102 +18,35 @@ use Psr\Http\Message\UriFactoryInterface;
 class Service extends AbstractService
 {
     /**
-     * @var UriFactoryInterface The PSR-17 URI factory instance.
+     * @var ServerRequestFactoryInterface The PSR-17 server request factory instance.
      */
-    protected $uriFactory;
+    protected $serverRequestFactory;
 
     /**
-     * @var StreamFactoryInterface The PSR-17 stream factory instance.
-     */
-    protected $streamFactory;
-
-    /**
-     * @var RequestFactoryInterface The PSR-17 request factory instance.
-     */
-    protected $requestFactory;
-
-    /**
-     * @var ResponseFactoryInterface The PSR-17 response factory instance.
-     */
-    protected $responseFactory;
-
-    /**
-     * @var ClientInterface The HTTP client instance.
-     */
-    protected $client;
-
-    /**
-     * Constructs the HTTP service with required factories.
+     * Constructs the HTTP service.
      *
-     * @param RequestFactoryInterface $requestFactory The request factory instance.
-     * @param ResponseFactoryInterface $responseFactory The response factory instance.
-     * @param StreamFactoryInterface $streamFactory The stream factory instance.
-     * @param ClientInterface $client The HTTP client instance.
+     * @param ServerRequestFactoryInterface $serverRequestFactory The server request factory.
      * @inheritDoc
      */
     public function __construct(
         ContainerInterface $container,
-        UriFactoryInterface $uriFactory,
-        StreamFactoryInterface $streamFactory,
-        RequestFactoryInterface $requestFactory,
-        ResponseFactoryInterface $responseFactory,
-        ClientInterface $client
+        ServerRequestFactoryInterface $serverRequestFactory
     ) {
         parent::__construct($container);
-        $this->uriFactory = $uriFactory;
-        $this->streamFactory = $streamFactory;
-        $this->requestFactory = $requestFactory;
-        $this->responseFactory = $responseFactory;
-        $this->client = $client;
+        $this->serverRequestFactory = $serverRequestFactory;
     }
 
     /**
-     * Returns the PSR-17 URI factory.
+     * Create a server request from the current environment.
      *
-     * @return UriFactoryInterface The request factory instance.
+     * @return ServerRequestInterface The created server request.
      */
-    public function getUriFactory(): UriFactoryInterface
+    public function createServerRequest(): ServerRequestInterface
     {
-        return $this->uriFactory;
-    }
-
-    /**
-     * Returns the PSR-17 stream factory.
-     *
-     * @return StreamFactoryInterface The stream factory instance.
-     */
-    public function getStreamFactory(): StreamFactoryInterface
-    {
-        return $this->streamFactory;
-    }
-
-    /**
-     * Returns the PSR-17 request factory.
-     *
-     * @return RequestFactoryInterface The request factory instance.
-     */
-    public function getRequestFactory(): RequestFactoryInterface
-    {
-        return $this->requestFactory;
-    }
-
-    /**
-     * Returns the PSR-17 response factory.
-     *
-     * @return ResponseFactoryInterface The response factory instance.
-     */
-    public function getResponseFactory(): ResponseFactoryInterface
-    {
-        return $this->responseFactory;
-    }
-
-    /**
-     * Returns the HTTP client.
-     *
-     * @return ClientInterface The HTTP client instance.
-     */
-    public function getClient(): ClientInterface
-    {
-        return $this->client;
+        return $this->serverRequestFactory->createServerRequest(
+            $_SERVER['REQUEST_METHOD'],
+            $_SERVER['REQUEST_URI'],
+            $_SERVER
+        );
     }
 }